Adjust AntoxFriend due to changes in jToxcore
Also slightly change the way Antox should be built locally.
This adds a link to the current "reference" build of jToxcore to
README.md. The dependencies should be downloaded from that build and
copied to the respective directories before building.

While this adds isTyping() functionality to the AntoxFriend class, it
does NOT implement any callbacks or complete functionality for Typing
notifications in Antox. These need to be implemented by the Antox team