Fix avatar not updating in navigation drawer view
BitmapManager never invalidated our own avatars, only friend avatars. This caused our avatar to be kept cached even when it was changed,  which in turn caused the navigation drawer view to only update on restart when the cache was cleared. Our own avatar is now invalidated when it is set in AvatarDialog.